// services/notifications/src/email/email.enum.ts
export enum TemplateType {
    Otp = "otp",
    WelcomeMessage = "welcomeMessage",
    AccountSuspended = "accountSuspended",
    AccountVerified = "accountVerified",
    MissedPayments = "missedPayments",
    PropertyAllocation = "propertyAllocation",
    ResetPassword = "resetPassword",
    UpdatedTermsAndConditions = "updatedTermsAndConditions",
    VerifyEmail = "verifyEmail",
    WalletTopUp = "walletTopUp",
    AdminContributionReceived = "adminContributionReceived",
    AdminPropertyAllocation = "adminPropertyAllocation",
    AdminInviteAdmin = "adminInviteAdmin",
}

export enum Section {
    DEVELOPER = 'developer',
    PMB = 'pmb',
    ADMIN = 'admin',
    BASE = 'base',
    MORTGAGE = 'mortgage',
    RSA = 'rsaApplication',
    PRELUANCH = 'preLaunch',
    CONTRIBUTION = 'contribution'
}